Move State into Migration namespace
